updated CollectionDownloadProgress to collectionDownloadProgress
